+++ SmartRegionManager.lua
@@ -5,7 +5,7 @@
   with automatic channel configuration based on region settings.
   
   Author: Unicon
-  Version: 1.1.1
+  Version: 1.1.2
   Requires: Reaper 6.0+, ReaImGui extension
 --]]